Calculations are WORKING!!! (Read Descriptioin)
They are working, HOWEVER, we do need timer delays on the function because the math wants to be faster than the updating state. The only issue with this is that if someone spams the button faster than the timer (50ms) then it will mess with the total and show incorrect data.